chore(ToolBar): rename FormatBar to ToolBar
- change Box to Divider
- set up flex using stack component
- remove TooltipToggleButton style margin and minWidth
